- removed some more bugs from the kernels, and the results are now closer to the CPU version.
- curand version are very fast for paths <= 1M, and slows down terribly
after that. On the other hand, performance of normrand versions improve
linearly upto ~33x.

- streamlined the output view to comapare CPU and GPU results

- Known bug: in curand_v2(), as volatility increases and/or as price at
t=0 gets close to strike-price, an upward bias is somehow added to the
estimated value of european option using MC. This bug is not present in
normrand_v1(), normrand_v2() and curand_v1(). Note that estimated value
of american options is unaffected by this bug. This should not be a
serious issue as closed form of european option could be used instead.